feat: add customization and filtering options for finding unused resources (#116)
* feat: add testify package

* feat: add the new FilterOptions struct

* chore: add optional FilterOptions to existing APIs

* feat: add the new persistent flag to root command

* fixup: replace wrong flag

* refactor: use newer-than and older-than flags for age filter

This commit changes the implementation of the HasIncludedAge function to use the newer-than and older-than flags instead of the MinAge and MaxAge flags. This makes the function more user-friendly and expressive, as users can specify queries like older-than 1h or newer-than 5m. The commit also updates the FilterOptions struct, the Validate function, and the flag descriptions accordingly.

* refactor: remove max-size and min-size flags

This commit removes max-size and min-size flags from filterOptions
flagset as well as all the size calculation functions defined in
resource_size.go. we may add them in the future.

* chore: update README file